# src/main/java/com/example/oauth2proxy/swagger/api/SignupApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.oauth2proxy.swagger.api;

import com.example.oauth2proxy.swagger.model.Error;
import com.example.oauth2proxy.swagger.model.SignUpRequest;
import com.example.oauth2proxy.swagger.model.SuccessOperationResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "Signup", description = "the Signup API")
public interface SignupApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /signup : Sign Up
     *
     * @param signUpRequest Sign Up Request (required)
     * @return Request served successfully, returning result (status code 200)
     *         or System Internal Error. Please contact System Administrator (status code 200)
     */
    @Operation(
        operationId = "registerUser",
        summary = "Sign Up",
        tags = { "Signup" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Request served successfully, returning result", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SuccessOperationResponse.class)),
                @Content(mediaType = "application/vnd.error+json", schema = @Schema(implementation = SuccessOperationResponse.class))
            }),
            @ApiResponse(responseCode = "default", description = "System Internal Error. Please contact System Administrator", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/vnd.error+json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/signup",
        produces = { "application/json", "application/vnd.error+json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<SuccessOperationResponse> registerUser(
        @Parameter(name = "SignUpRequest", description = "Sign Up Request", required = true) @Valid @RequestBody SignUpRequest signUpRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"result\" : \"result\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/vnd.error+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/vnd.error+json";
                    ApiUtil.setExampleResponse(request, "application/vnd.error+json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}